pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tourbox-macos"
version = "0.1.0"
description = "Custom macOS driver for TourBox Lite with AI-powered profile generation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Gavin" }
]
keywords = ["tourbox", "macos", "driver", "keyboard", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware :: Hardware Drivers",
]

dependencies = [
    "pyserial>=3.5",
    "pynput>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
tourbox = "tourbox_macos.driver:main"

[project.urls]
Homepage = "https://github.com/goov373/tourbox-macos"
Repository = "https://github.com/goov373/tourbox-macos"
Issues = "https://github.com/goov373/tourbox-macos/issues"

# Ruff - Fast Python linter and formatter
[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

# MyPy - Type checking
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

# Hatch - Build configuration
[tool.hatch.build.targets.wheel]
packages = ["src/tourbox_macos"]
